+++ src/pact_consumer_http.erl
@@ -79,23 +79,38 @@ start_mock_server(PactPid, PactRef, Host, Port, InteractionPart) ->
 
 -spec insert_request_details(pact_interaction_ref(), request_details()) -> ok.
 insert_request_details(InteractionRef, RequestDetails) ->
+    %% Checking if someone used regex_match
+    CheckIfMap =
+        fun(Value) ->
+            case is_map(Value) of
+                true ->
+                    thoas:encode(Value);
+                false ->
+                    Value
+            end
+        end,
     ReqMethod = maps:get(method, RequestDetails),
     ReqPath = maps:get(path, RequestDetails),
-    pactffi_nif:with_request(InteractionRef, ReqMethod, ReqPath),
+    NewReqPath = CheckIfMap(ReqPath),
+    pactffi_nif:with_request(InteractionRef, ReqMethod, NewReqPath),
     ReqHeaders = maps:get(headers, RequestDetails, #{}),
     ContentType = get_content_type(ReqHeaders),
     maps:fold(
         fun(Key, Value, _Acc) ->
             %% FIXME: 4th parameter is Index.. need to increment
-            pactffi_nif:with_header_v2(InteractionRef, 0, Key, 0, Value)
+            NewValue = CheckIfMap(Value),
+            pactffi_nif:with_header_v2(InteractionRef, 0, Key, 0, NewValue)
         end,
         ok,
         ReqHeaders
     ),
     ReqBody = maps:get(body, RequestDetails, undefined),
     case ReqBody of
-        undefined -> ok;
-        _ -> pactffi_nif:with_body(InteractionRef, 0, ContentType, ReqBody)
+        undefined ->
+            ok;
+        _ ->
+            NewReqBody = CheckIfMap(ReqBody),
+            pactffi_nif:with_body(InteractionRef, 0, ContentType, NewReqBody)
     end,
     ReqQueryParams = maps:get(query_params, RequestDetails, undefined),
     case ReqQueryParams of
@@ -105,7 +120,8 @@ insert_request_details(InteractionRef, RequestDetails) ->
             maps:fold(
                 fun(Key, Value, _Acc) ->
                     %% FIXME: 3rd parameter is Index.. need to increment
-                    pactffi_nif:with_query_parameter_v2(InteractionRef, Key, 0, Value)
+                    NewValue = CheckIfMap(Value),
+                    pactffi_nif:with_query_parameter_v2(InteractionRef, Key, 0, NewValue)
                 end,
                 ok,
                 ReqQueryParams
@@ -115,6 +131,16 @@ insert_request_details(InteractionRef, RequestDetails) ->
 
 -spec insert_response_details(pact_interaction_ref(), response_details()) -> ok.
 insert_response_details(InteractionRef, ResponseDetails) ->
+    %% Checking if someone used regex_match
+    CheckIfMap =
+        fun(Value) ->
+            case is_map(Value) of
+                true ->
+                    thoas:encode(Value);
+                false ->
+                    Value
+            end
+        end,
     ResponseStatusCode = maps:get(status, ResponseDetails, undefined),
     case ResponseStatusCode of
         undefined -> ok;
@@ -125,15 +151,19 @@ insert_response_details(InteractionRef, ResponseDetails) ->
     maps:fold(
         fun(Key, Value, _Acc) ->
             %% FIXME: 4th parameter is Index.. need to increment
-            pactffi_nif:with_header_v2(InteractionRef, 1, Key, 0, Value)
+            NewValue = CheckIfMap(Value),
+            pactffi_nif:with_header_v2(InteractionRef, 1, Key, 0, NewValue)
         end,
         ok,
         ResHeaders
     ),
     ResBody = maps:get(body, ResponseDetails, undefined),
     case ResBody of
-        undefined -> ok;
-        _ -> pactffi_nif:with_body(InteractionRef, 1, ContentType, ResBody)
+        undefined ->
+            ok;
+        _ ->
+            NewResBody = CheckIfMap(ResBody),
+            pactffi_nif:with_body(InteractionRef, 1, ContentType, NewResBody)
     end,
     ok.