+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Introduced ability to add params inside given in pact interaction
+- Introduced basic pact matchers
+
 
 ## [0.2.2] - 2023-11-03